Chapter 5 complete source (presumed).
Created repositories for the entities.
Created base repository class to dispose of the context.
Created view models for the views.
Moved PagedResult views into the ViewModel folder.
Created new abstraction for PagedResult to specify type.
Created views and actions to create, display, and edit recipes.
Changed navigation to say Recipes instead of Home.
Modified recipe controller to dispose of repositories.
Removed several unused using statements.
Removed several Model. and Entity. calls as they're not necessary.

This check-in is large because it's the complete commit for a chapter.
This is NOT good practice.